requirements.txt
PyQt5==5.15.7
psutil==5.9.4
pygetwindow==0.0.9
pyinstaller==6.14.1
pillow==10.2.0 